requirements.txt
beautifulsoup4==4.10.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
idna==3.3
importlib-metadata==4.10.1
jieba3k==0.35.1
joblib==1.1.0
lxml==4.7.1
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.6.7
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
shortuuid==1.0.8
six==1.16.0
soupsieve==2.3.1
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.1.2
tomli==1.2.3
tqdm==4.62.3
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
